# packages/contracts/openapi/collections-v1.yaml
openapi: 3.1.0
info:
  title: DeepDocParse Explicit Collection Catalog
  version: 1.0.0
  description: See ../ddp/collection-catalog-format.md. Corpus owns collections and all source authorization.
security: [{session: []}, {userKey: []}]
paths:
...
  /internal/federation/collections:
    get:
      operationId: enumeratePublishedCollections
      security: [{service: []}]
      description: Requires service actor and all trusted caller headers documented in collection-catalog-format.md. Never directly proxy user requests to this path.
      parameters:
        - {name: scope_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
        - {name: snapshot_id, in: query, schema: {type: string, maxLength: 32}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 32}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Stable authorized metadata page with separate empty terminal proof, content: {application/json: {schema: {$ref: '#/components/schemas/CatalogPage'}}}}
        '401': {description: Service or trusted caller context missing}
        '403': {description: Non-service actor}
        '409': {description: catalog_snapshot_changed; index/member revision changed, not a revocation}
        '410': {description: catalog_snapshot_expired or catalog_snapshot_invalid. A bound unexpired owner may also receive revoked_collection_ids, restricted to original snapshot identities.}
        '429': {description: Caller snapshot storage bound reached}
        '507': {description: Authorized catalog too large; never claim complete}
...
components:
  securitySchemes:
    session: {type: http, scheme: bearer}
    userKey: {type: apiKey, in: header, name: X-API-Key}
    service: {type: http, scheme: bearer}
    nodeCredential: {type: apiKey, in: header, name: X-DDP-Node-Credential, description: Single-use Ed25519 directory credential with live approval and durable replay protection}
...
  schemas:
...
    CatalogPage:
      type: object
      properties:
        snapshot_id: {type: string, pattern: '^[0-9a-f]{32}$'}
        scope_id: {type: string, minLength: 1, maxLength: 128}
        caller_scope_hash: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        origin_node_id: {$ref: '../schemas/ddp-discovery/v1.json#/$defs/node_id'}
        registry_revision: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        first_cursor: {type: string, pattern: '^[0-9a-f]{32}$'}
        terminal_cursor: {type: string, pattern: '^[0-9a-f]{32}$'}
        total: {type: integer, minimum: 0, maximum: 10000}
        collections: {type: array, maxItems: 100, items: {$ref: '../schemas/ddp-discovery/v1.json#/$defs/CollectionDescriptor'}}
        index_readiness: {type: object, additionalProperties: {type: string, enum: [ready, unavailable]}}
        next_cursor: {type: [string, 'null'], pattern: '^[0-9a-f]{32}$'}
        complete: {type: boolean}
        content_snapshot_complete: {const: false}
      required: [snapshot_id, scope_id, caller_scope_hash, origin_node_id, registry_revision, created_at, valid_until, first_cursor, terminal_cursor, total, collections, index_readiness, next_cursor, complete, content_snapshot_complete]
      additionalProperties: false
      allOf:
        - if: {properties: {complete: {const: true}}}
          then:
            properties: {collections: {maxItems: 0}, next_cursor: {const: null}, index_readiness: {maxProperties: 0}}
...